# [LOGCXX-546] Prevent unnecessary locking when determining the logging level