--- a/src/test/cpp/customlogger/xlogger.cpp
+++ b/src/test/cpp/customlogger/xlogger.cpp
@@ -32,8 +32,8 @@ XFactoryPtr XLogger::factory = XFactoryPtr(new XFactory());
 
 void XLogger::lethal(const LogString& message, const LocationInfo& locationInfo)
 {
-	log4cxx::spi::LoggerRepositoryPtr rep = repository.lock();
-	if (rep->isDisabled(XLevel::LETHAL_INT))
+	auto rep = repository;
+	if (!rep || rep->isDisabled(XLevel::LETHAL_INT))
 	{
 		return;
 	}
@@ -46,8 +46,8 @@ void XLogger::lethal(const LogString& message, const LocationInfo& locationInfo)
 
 void XLogger::lethal(const LogString& message)
 {
-	log4cxx::spi::LoggerRepositoryPtr rep = repository.lock();
-	if (rep->isDisabled(XLevel::LETHAL_INT))
+	auto rep = repository;
+	if (!rep || rep->isDisabled(XLevel::LETHAL_INT))
 	{
 		return;
 	}
@@ -70,8 +70,8 @@ LoggerPtr XLogger::getLogger(const helpers::Class& clazz)
 
 void XLogger::trace(const LogString& message, const LocationInfo& locationInfo)
 {
-	log4cxx::spi::LoggerRepositoryPtr rep = repository.lock();
-	if (rep->isDisabled(XLevel::TRACE_INT))
+	auto rep = repository;
+	if (!rep || rep->isDisabled(XLevel::TRACE_INT))
 	{
 		return;
 	}
@@ -84,8 +84,8 @@ void XLogger::trace(const LogString& message, const LocationInfo& locationInfo)
 
 void XLogger::trace(const LogString& message)
 {
-	log4cxx::spi::LoggerRepositoryPtr rep = repository.lock();
-	if (rep->isDisabled(XLevel::TRACE_INT))
+	auto rep = repository;
+	if (!rep || rep->isDisabled(XLevel::TRACE_INT))
 	{
 		return;
 	}