--- a/src/test/resources/witness/autoConfigureTest.log
+++ b/src/test/resources/witness/autoConfigureTest.log
@@ -1,0 +1,1 @@
+This is some expected ouput