**throughputtests after these changes:**
```
Benchmarking library only(no writing out):
**************************************************************
Benchmarking Single threaded: 1000000 messages
**************************************************************
Log4cxx NoFormat pattern: %m%n Elapsed: 2.658 secs 376,183/sec
Log4cxx DateOnly pattern: [%d] %m%n Elapsed: 2.665 secs 375,241/sec
Log4cxx DateClassLevel pattern: [%d] [%c] [%p] %m%n Elapsed: 2.688 secs 372,005/sec
Log4cxx Logging with FMT Elapsed: 2.137 secs 468,036/sec
Log4cxx Logging static string Elapsed: 2.143 secs 466,650/sec
Log4cxx Logging static string with FMT Elapsed: 2.245 secs 445,344/sec
Log4cxx Logging disabled debug Elapsed: 0.03112 secs 32,134,432/sec
Log4cxx Logging disabled trace Elapsed: 0.03091 secs 32,355,477/sec
Log4cxx Logging enabled debug Elapsed: 2.142 secs 466,769/sec
Log4cxx Logging enabled trace Elapsed: 2.161 secs 462,705/sec
**************************************************************
Benchmarking multithreaded threaded: 1000000 messages/thread, 4 threads
**************************************************************
Log4cxx Logging with FMT MT Elapsed: 4.19 secs 238,661/sec
Log4cxx Logging with FMT MT Elapsed: 4.233 secs 236,256/sec
Log4cxx Logging with FMT MT Elapsed: 4.315 secs 231,733/sec
Log4cxx Logging with FMT MT Elapsed: 4.324 secs 231,272/sec
**************************************************************
Benchmarking multithreaded disabled: 1000000 messages/thread, 4 threads
**************************************************************
Log4cxx Logging disabled MT Elapsed: 0.03375 secs 29,632,945/sec
Log4cxx Logging disabled MT Elapsed: 0.03366 secs 29,706,781/sec
Log4cxx Logging disabled MT Elapsed: 0.03372 secs 29,653,837/sec
Log4cxx Logging disabled MT Elapsed: 0.03375 secs 29,627,664/sec
```

**throughput output before these changes:**
```**************************************************************
Benchmarking Single threaded: 1000000 messages
**************************************************************
Log4cxx NoFormat pattern: %m%n Elapsed: 2.828 secs 353,625/sec
Log4cxx DateOnly pattern: [%d] %m%n Elapsed: 2.767 secs 361,350/sec
Log4cxx DateClassLevel pattern: [%d] [%c] [%p] %m%n Elapsed: 2.81 secs 355,882/sec
Log4cxx Logging with FMT Elapsed: 2.088 secs 478,949/sec
Log4cxx Logging static string Elapsed: 2.217 secs 451,045/sec
Log4cxx Logging static string with FMT Elapsed: 2.302 secs 434,344/sec
Log4cxx Logging disabled debug Elapsed: 0.1867 secs 5,357,277/sec
Log4cxx Logging disabled trace Elapsed: 0.186 secs 5,375,506/sec
Log4cxx Logging enabled debug Elapsed: 2.261 secs 442,245/sec
Log4cxx Logging enabled trace Elapsed: 2.263 secs 441,886/sec
**************************************************************
Benchmarking multithreaded threaded: 1000000 messages/thread, 4 threads
**************************************************************
Log4cxx Logging with FMT MT Elapsed: 4.097 secs 244,063/sec
Log4cxx Logging with FMT MT Elapsed: 4.105 secs 243,617/sec
Log4cxx Logging with FMT MT Elapsed: 4.132 secs 242,014/sec
Log4cxx Logging with FMT MT Elapsed: 4.132 secs 242,005/sec
**************************************************************
Benchmarking multithreaded disabled: 1000000 messages/thread, 4 threads
**************************************************************
Log4cxx Logging disabled MT Elapsed: 0.6538 secs 1,529,611/sec
Log4cxx Logging disabled MT Elapsed: 0.6681 secs 1,496,778/sec
Log4cxx Logging disabled MT Elapsed: 0.6699 secs 1,492,832/sec
Log4cxx Logging disabled MT Elapsed: 0.6702 secs 1,492,142/sec
```