@rm5248 2021-12-29
Looks good to me.  I also get rather dramatic increases in the number of log messages/second when logging is disabled(approx. 10x on my machine).  Other stats are about 5-10% better.