@swebb2066 2021-12-29
Do you think this performance fix warrent a release (0.12.3) before next_stable is ready?