@rm5248 2021-12-29
My current thought is that we'll do a 0.13.0 release in a few weeks; there are a number of fixes for that already in master.  next_stable probably won't be for several months at least, since that contains bigger changes(e.g. API breaking, other assorted fixes that are currently hard to fix with the current state of the code without touching pretty much everything).