@ams-tschoening 2021-12-30
Before:

```text
Benchmarking library only(no writing out):
**************************************************************
Benchmarking Single threaded: 1000000 messages
**************************************************************
Log4cxx NoFormat pattern: %m%n Elapsed: 30.14 secs 33.174/sec
Log4cxx DateOnly pattern: [%d] %m%n Elapsed: 24.65 secs 40.569/sec
Log4cxx DateClassLevel pattern: [%d] [%c] [%p] %m%n Elapsed: 22.96 secs 43.549/sec
Log4cxx Logging with FMT Elapsed: 23.98 secs 41.701/sec
Log4cxx Logging static string Elapsed: 9.429 secs 106.055/sec
Log4cxx Logging static string with FMT Elapsed: 28.13 secs 35.551/sec
Log4cxx Logging disabled debug Elapsed: 0.24 secs 4.166.673/sec
Log4cxx Logging disabled trace Elapsed: 0.2271 secs 4.403.643/sec
Log4cxx Logging enabled debug Elapsed: 8.44 secs 118.484/sec
Log4cxx Logging enabled trace Elapsed: 8.734 secs 114.493/sec
**************************************************************
Benchmarking multithreaded threaded: 1000000 messages/thread, 4 threads
**************************************************************
Log4cxx Logging with FMT MT Elapsed: 143 secs 6.991/sec
Log4cxx Logging with FMT MT Elapsed: 143.1 secs 6.988/sec
Log4cxx Logging with FMT MT Elapsed: 143.2 secs 6.984/sec
Log4cxx Logging with FMT MT Elapsed: 143.2 secs 6.982/sec
**************************************************************
Benchmarking multithreaded disabled: 1000000 messages/thread, 4 threads
**************************************************************
Log4cxx Logging disabled MT Elapsed: 0.7581 secs 1.319.082/sec
Log4cxx Logging disabled MT Elapsed: 0.7681 secs 1.301.911/sec
Log4cxx Logging disabled MT Elapsed: 0.7699 secs 1.298.865/sec
Log4cxx Logging disabled MT Elapsed: 0.775 secs 1.290.337/sec
```

After:

```text
Benchmarking library only(no writing out):
**************************************************************
Benchmarking Single threaded: 1000000 messages
**************************************************************
Log4cxx NoFormat pattern: %m%n Elapsed: 30.23 secs 33.075/sec
Log4cxx DateOnly pattern: [%d] %m%n Elapsed: 19.43 secs 51.465/sec
Log4cxx DateClassLevel pattern: [%d] [%c] [%p] %m%n Elapsed: 18.13 secs 55.144/sec
Log4cxx Logging with FMT Elapsed: 21.12 secs 47.337/sec
Log4cxx Logging static string Elapsed: 7.319 secs 136.623/sec
Log4cxx Logging static string with FMT Elapsed: 24.16 secs 41.388/sec
Log4cxx Logging disabled debug Elapsed: 0.04665 secs 21.436.181/sec
Log4cxx Logging disabled trace Elapsed: 0.04477 secs 22.333.941/sec
Log4cxx Logging enabled debug Elapsed: 7.426 secs 134.654/sec
Log4cxx Logging enabled trace Elapsed: 7.309 secs 136.813/sec
**************************************************************
Benchmarking multithreaded threaded: 1000000 messages/thread, 4 threads
**************************************************************
Log4cxx Logging with FMT MT Elapsed: 140.7 secs 7.105/sec
Log4cxx Logging with FMT MT Elapsed: 141.1 secs 7.088/sec
Log4cxx Logging with FMT MT Elapsed: 141.2 secs 7.081/sec
Log4cxx Logging with FMT MT Elapsed: 141.3 secs 7.078/sec
**************************************************************
Benchmarking multithreaded disabled: 1000000 messages/thread, 4 threads
**************************************************************
Log4cxx Logging disabled MT Elapsed: 0.03834 secs 26.083.917/sec
Log4cxx Logging disabled MT Elapsed: 0.04217 secs 23.711.966/sec
Log4cxx Logging disabled MT Elapsed: 0.0465 secs 21.504.821/sec
Log4cxx Logging disabled MT Elapsed: 0.05066 secs 19.740.686/sec
```